added consumer commit logic

+++ AsyncQueue/Domain/Models/ConsumersDtos/ConsumerPollResponse.cs
@@ -5,7 +5,7 @@ public class ConsumerPollResponse
     public ProcessingStatus ProcessingStatus { get; set; }
     public string? Message { get; set; }
     public int PartitionId { get; set; }
-    public long Offset { get; set; }
+    public int Offset { get; set; }
     public string? ValueType { get; set; }
     public List<ConsumerMessage>? ConsumerMessages { get; set; }
 }
